requirements.txt
# Dependências principais do projeto
ultralytics>=8.0.0
opencv-python>=4.8.0
Pillow>=9.0.0
numpy>=1.21.0
torch>=2.0.0
torchvision>=0.15.0

# Tracking e análise de objetos
supervision>=0.16.0
scipy>=1.9.0

# Manipulação de dados
pandas>=1.5.0
matplotlib>=3.6.0

# Utilitários
pathlib2>=2.3.0
tqdm>=4.64.0

# Opcional: Para tracking mais avançado (se necessário)
# filterpy>=1.4.5
# lap>=0.4.0